Moved content from view to presenter
