+++ client/src/components/views/createView.tsx
@@ -1,116 +1,35 @@
 import "../../assets/styles/create.scss";
-import { useState } from "react";
 
 interface CreateViewProps {
   userMessage: string;
   setUserMessage: (message: string) => void;
   botResponse: string;
-  setBotResponse: (response: string) => void;
   onSubmit: () => void;
+  code: string;
+  token: string;
+  email: string;
+  userName: string;
+  onLoad: () => void;
+  onLogin: () => void;
+  onToken: () => void;
+  onProfile: () => void;
 }
-
-const params = new URLSearchParams(window.location.search);
-const code = params.get("code");
 
 const CreateView: React.FC<CreateViewProps> = ({
   userMessage,
   setUserMessage,
   botResponse,
   onSubmit,
+  code,
+  token,
+  email,
+  userName,
+  onLoad,
+  onLogin,
+  onToken,
+  onProfile,
 }) => {
-  const [userName, setUserName] = useState<string>("placeholder");
-  const [token, setToken] = useState<string>("placeholder");
-  const [profileInfo, setProfileInfo] = useState<string>("placeholder");
-
-  const handleSubmitAPI = async () => {
-    function generateCodeVerifier(length: number) {
-      let text = "";
-      let possible =
-        "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-
-      for (let i = 0; i < length; i++) {
-        text += possible.charAt(Math.floor(Math.random() * possible.length));
-      }
-      return text;
-    }
-
-    async function generateCodeChallenge(codeVerifier: string) {
-      const data = new TextEncoder().encode(codeVerifier);
-      const digest = await window.crypto.subtle.digest("SHA-256", data);
-      return btoa(String.fromCharCode.apply(null, [...new Uint8Array(digest)]))
-        .replace(/\+/g, "-")
-        .replace(/\//g, "_")
-        .replace(/=+$/, "");
-    }
-
-    try {
-      const verifier = generateCodeVerifier(128);
-      const challenge = await generateCodeChallenge(verifier);
-
-      localStorage.setItem("verifier", verifier);
-      const apiUrl = `http://localhost:3001/api/login?challenge=${encodeURIComponent(
-        challenge
-      )}`;
-      const response = await fetch(apiUrl, {
-        method: "GET",
-      });
-
-      if (!response.ok) {
-        throw new Error(`HTTP error! Status: ${response.status}`);
-      }
-
-      const data = await response.json();
-      document.location = data.urlResponse;
-    } catch (error) {
-      console.error("Error:", error);
-    }
-  };
-
-  const getToken = async () => {
-    const verifier = localStorage.getItem("verifier");
-    if (typeof code === "string" && typeof verifier === "string") {
-      try {
-        const response = await fetch(`http://localhost:3001/api/token`, {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({ code: code, verifier: verifier }),
-        });
-
-        if (!response.ok) {
-          throw new Error(`HTTP error! Status: ${response.status}`);
-        }
-
-        const data = await response.json();
-        setToken(data.token);
-      } catch (error) {
-        console.error("Error:", error);
-      }
-    }
-  };
-
-  const getProfile = async () => {
-    try {
-
-      const response = await fetch(
-        `http://localhost:3001/api/profile?token=${encodeURIComponent(token)}`,
-        {
-          method: "GET",
-        }
-      );
-
-      if (!response.ok) {
-        throw new Error(`HTTP error! Status: ${response.status}`);
-      }
-
-      const data = await response.json();
-      setProfileInfo(data.profile.email);
-      setUserName(data.profile.display_name);
-    } catch (error) {
-      console.error("Error:", error);
-    }
-  };
+  onLoad();
 
   const handleSubmit = (e: React.FormEvent) => {
     e.preventDefault();
@@ -139,12 +58,12 @@ const CreateView: React.FC<CreateViewProps> = ({
           <p>{code}</p>
           <p>Token:</p>
           <p>{token}</p>
-          <p>Profile information: </p>
-          <p>{profileInfo}</p>
+          <p>Email: </p>
+          <p>{email}</p>
         </div>
-        <button onClick={handleSubmitAPI}>Test API</button>
-        <button onClick={getToken}>Get token</button>
-        <button onClick={getProfile}>Get profile information</button>
+        <button onClick={onLogin}>Test API</button>
+        <button onClick={onToken}>Get token</button>
+        <button onClick={onProfile}>Get profile information</button>
       </div>
     </div>
   );
